+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.27.6 (2024-06-18)
+
+### Bug Fixes and Other Changes
+
+ * Support `qml.counts` for circuits
+
+## v1.27.5 (2024-06-13)
+
+### Bug Fixes and Other Changes
+
+ * Replace `FreeParameter("b")` with `FreeParameter("d")` in unittests
+
 ## v1.27.4 (2024-06-03)
 
 ### Bug Fixes and Other Changes